chilli_leaf_ai_2/requirements.txt
tensorflow-cpu
keras
opencv-python-headless
numpy
matplotlib
scikit-learn
pillow
flask
tqdm
gunicorn
ultralytics
